picmi: react to review on particle region binning
- single source of truth: PARTICLE_REGIONS + derived ParticleRegion
  defined once in pypicongpu output/binning.py, imported by picmi
  diagnostics (drop duplicated literal list).
- drop dedicated test_binning.py per review; fold the crucial
  particle-region assertions into quick/picmi/test_simulation.py.
